refactor(session): trim drizzle schema surface
