+++ packages/session/src/storage/drizzle/db.ts
@@ -1,9 +1,9 @@
 import { Database } from "bun:sqlite";
 import { drizzle } from "drizzle-orm/bun-sqlite";
-import * as schema from "./schema";
 import { initializeSqliteDatabase } from "../sqlite-schema-lifecycle";
+import { drizzleSchema } from "./schema";
 
-export type DrizzleDb = ReturnType<typeof drizzle<typeof schema>>;
+export type DrizzleDb = ReturnType<typeof drizzle<typeof drizzleSchema>>;
 
 export function createDb(dbPath: string): { db: DrizzleDb; sqlite: Database } {
   const sqlite = new Database(dbPath);
@@ -11,7 +11,7 @@ export function createDb(dbPath: string): { db: DrizzleDb; sqlite: Database } {
   try {
     initializeSqliteDatabase(sqlite);
 
-    const db = drizzle(sqlite, { schema });
+    const db = drizzle(sqlite, { schema: drizzleSchema });
     return { db, sqlite };
   } catch (err) {
     sqlite.close();
